replase helpers and config, fix validation
