+++ product-service/products/get-product-list.ts
@@ -1,17 +1,7 @@
 import * as AWSLambda from "aws-lambda";
 import { CORS_HEADERS } from "../cors-headers";
-import { Client, ClientConfig } from "pg";
-
-const { DB_HOST, DB_PORT, DB_DATABASE, DB_USERNAME, DB_PASSWORD } = process.env;
-const dbOptions: ClientConfig = {
-  host: DB_HOST,
-  port: +DB_PORT,
-  database: DB_DATABASE,
-  user: DB_USERNAME,
-  password: DB_PASSWORD,
-  ssl: { rejectUnauthorized: false },
-  connectionTimeoutMillis: 5000,
-};
+import { Client } from "pg";
+import { dbOptions } from "./helper/db-config";
 
 export const getProductList = async (event: AWSLambda.APIGatewayEvent) => {
   const client = new Client(dbOptions);
