+++ product-service/products/helper/api-error.ts
@@ -1,0 +1,8 @@
+export class ApiError extends Error {
+  private statusCode: number;
+  constructor(name: string, statusCode: number, message?: string) {
+    super(message);
+    this.name = name;
+    this.statusCode = statusCode;
+  }
+}
