+++ product-service/products/helper/db-config.ts
@@ -1,0 +1,12 @@
+import { ClientConfig } from "pg";
+
+const { DB_HOST, DB_PORT, DB_DATABASE, DB_USERNAME, DB_PASSWORD } = process.env;
+export const dbOptions: ClientConfig = {
+  host: DB_HOST,
+  port: +DB_PORT,
+  database: DB_DATABASE,
+  user: DB_USERNAME,
+  password: DB_PASSWORD,
+  ssl: { rejectUnauthorized: false },
+  connectionTimeoutMillis: 5000,
+};
